fix(flow_solver.c): fix proximity check
Proximity check was previously done after taking the square root. This would be an issue when the control point was almost in the same position as one of the line's endpoints and the normal distance would be -eps. By moving the proximity check before the square root, this now correctly gets recognized as too close and no square root happens.